pip_requirements_frozen.txt
absl-py==2.1.0
accelerate==0.21.0
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
autogluon==1.1.1
autogluon.common==1.1.1
autogluon.core==1.1.1
autogluon.features==1.1.1
autogluon.multimodal==1.1.1
autogluon.tabular==1.1.1
autogluon.timeseries==1.1.1
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
boto3==1.34.133
botocore==1.34.133
cachetools==5.3.3
catalogue==2.0.10
catboost==1.2.5
category-encoders==2.6.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
colorlog==6.8.2
comm @ file:///croot/comm_1709322850197/work
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
datasets==2.20.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
evaluate==0.4.2
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastai==2.7.15
fastcore==1.5.48
fastdownload==0.0.7
fastjsonschema==2.20.0
fastprogress==1.0.3
filelock==3.15.4
flatbuffers==24.3.25
focal-loss==0.0.7
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gast==0.5.5
gdown==5.2.0
gluonts==0.15.1
google-api-core==2.19.1
google-auth==2.30.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
GRANDE==0.1.6
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperopt==0.2.7
idna==3.7
imageio==2.34.2
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1718287989724/work
isoduration==20.11.0
jedi @ file:///work/ci_py311_2/jedi_1679336495545/work
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.6.17
keras==3.4.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
liac-arff==2.5.0
libclang==18.1.1
lightgbm==4.3.0
lightning==2.3.0
lightning-utilities==0.11.2
llvmlite==0.43.0
Mako==1.3.5
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mdurl==0.1.2
minio==7.2.7
mistune==3.0.2
ml-dtypes==0.3.2
ml-metrics==0.1.4
mlforecast==0.10.0
model-index==0.1.11
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.3
nlpaug==1.1.11
nltk==3.8.1
notebook_shim==0.2.4
nptyping==2.4.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.2.3
onnx==1.16.1
onnxruntime==1.18.0
opencensus==0.11.4
opencensus-context==0.1.3
opendatalab==0.0.10
openfe @ git+https://github.com/IIIS-Li-Group/OpenFE.git@3e3bdf9f5b972e214e018c22f6f9f456a47b416e
openmim==0.3.9
openml==0.14.2
openxlab==0.0.11
opt-einsum==3.3.0
optimum==1.18.1
optree==0.11.0
optuna==3.6.1
ordered-set==4.1.0
orjson==3.10.5
overrides==7.7.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
patsy==0.5.6
pdf2image==1.17.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==10.3.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
plotly==5.22.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
proto-plus==1.24.0
protobuf==4.25.3
psutil @ file:///work/ci_py311_2/psutil_1679337388738/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing==3.1.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-json-logger==2.0.7
python-slugify==8.0.4
pytorch-lightning==2.3.0
pytorch-metric-learning==2.3.0
pytorch_frame==0.2.2
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
ray==2.31.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
rtdl_num_embeddings==0.0.9
rtdl_revisiting_models==0.0.2
s3transfer==0.10.2
safetensors==0.4.3
scikit-image==0.20.0
scikit-learn==1.4.0
scipy==1.12.0
Send2Trash==1.8.3
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
statsforecast==1.4.0
statsmodels==0.14.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.6.18
timm==0.9.16
tinycss2==1.3.0
tokenizers==0.15.2
toolz==0.12.1
torch==2.2.1
torchcontrib==0.0.2
torcheval==0.0.7
torchmetrics==1.2.1
torchvision==0.17.1
tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.66.4
traitlets @ file:///croot/traitlets_1718227057033/work
transformers==4.39.3
triton==2.2.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
utilsforecast==0.0.10
virtualenv==20.26.3
wasabi==1.1.3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
window_ops==0.0.15
wrapt==1.16.0
xfeat==0.1.1
xgboost==2.0.3
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.4
